Broke out sendUrlPostRequest function to be imported into main.js. #2
