+++ client/main.js
@@ -1,37 +1,9 @@
 const createFormTable = require('./utils/createFormTable.js')
-const invalidUrlMessage = require('./utils/invalidUrlMessage.js')
-const sendTracklistPostRequest = require('./sendTracklistPostRequest.js')
+const sendTracklistPostRequest = require('./utils/sendTracklistPostRequest.js')
+const sendUrlPostRequest = require('./utils/sendUrlPostRequest.js')
 const submitTracklist = require('./utils/submitTracklist.js')
-const transitionToTracklistForm = require('./utils/transitionToTracklistForm.js')
 
 const HashRouter = require('./utils/hashRouter.js')
-
-function sendUrlPostRequest(urlSubmission) {
-  fetch('/url-request', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-      'Accept': 'application/json'
-    },
-    body: JSON.stringify(urlSubmission)
-  })
-  .then(response => {
-    console.log(response)
-    if (response.status === 400) {
-      const $invalid = invalidUrlMessage()
-      const $urlFormGroup = document.getElementById('url-form-col')
-      $urlFormGroup.appendChild($invalid)
-    }
-    if (response.status === 202) return response.json()
-  })
-  .then(keyData => {
-    window.location.hash = '#create-tracklist' + '?id=' + keyData.videoId
-    transitionToTracklistForm(keyData)
-    console.log(keyData)
-    albumMetadata = keyData
-  })
-  .catch(err => console.log(err))
-}
 
 function validateUrl(url) {
   return url.includes('https://www.youtube.com/')
@@ -59,7 +31,9 @@ $submitButton.addEventListener('click', () => {
   if (validateUrl($urlInput.value)) {
     urlSubmission.url = $urlInput.value
     urlSubmission.youtubeId = getYoutubeId(urlSubmission.url)
-    sendUrlPostRequest(urlSubmission)
+    sendUrlPostRequest(urlSubmission).then(keyData => {
+      albumMetadata = keyData
+    })
   }
 })
